+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 ## 06/10/2022 3.0.1
 
+- ModelValidator - Report real exampleJsonPath when additional parameter includes "."
 - Support duration format when generating examples
 - SemanticValidator - Fix bug about multipart/form-data